Try Catch Block Removed - Using Express Async Handler

+++ src/controllers/appointment-controller.ts
@@ -1,34 +1,38 @@
 import express, { Request, Response, NextFunction } from "express";
+import asyncHandler from "express-async-handler";
+import mongoose from "mongoose";
 import { IAppointments } from "../models/appointment";
 import AppointmentModel from "../models/appointment";
 import { IPatient } from "../models/patient";
 import PatientModel from "../models/patient";
 
-export const getAllAppointmentForPatient = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    const appointments: any = await AppointmentModel.find({
-      patientId: req.params.patId,
+export const getAllAppointmentForPatient = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { patId } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(patId)) {
+      return next({ statusCode: 400, message: "Invalid Patient ID format" });
+    }
+    const appointments = await AppointmentModel.find({
+      patientId: patId,
     }).populate("patientId");
-    if (!appointments) {
+    if (!appointments || appointments.length === 0) {
       return next({ statusCode: 404, message: "Appointments not found" });
     }
     res.status(200).json(appointments);
-  } catch (error) {
-    next({ statusCode: 500, message: "Internal server error" });
   }
-};
-export const updateAppointment = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+);
+
+export const updateAppointment = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { appointId } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(appointId)) {
+      return next({
+        statusCode: 400,
+        message: "Invalid Appointment ID format",
+      });
+    }
     const updatedPost = await AppointmentModel.findByIdAndUpdate(
-      req.params.appointId,
+      appointId,
       req.body,
       { new: true }
     ).populate("patientId");
@@ -36,77 +40,61 @@ export const updateAppointment = async (
       return next({ statusCode: 404, message: "Appointment not found" });
     }
     res.status(200).json(updatedPost);
-  } catch (error) {
-    next({ statusCode: 500, message: "Internal server error" });
   }
-};
-
-export const deleteAppointment = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+);
+export const deleteAppointment = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { appointId } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(appointId)) {
+      return next({
+        statusCode: 400,
+        message: "Invalid Appointment ID format",
+      });
+    }
     const deletedPost = await AppointmentModel.findByIdAndDelete(
-      req.params.appointId,
-      req.body
+      appointId
     ).populate("patientId");
     if (!deletedPost) {
       return next({ statusCode: 404, message: "Appointment not found" });
     }
     res.status(200).json({ message: "Appointment deleted successfully" });
-  } catch (error) {
-    //console.error("Error deleting appointment:", error);
-    return next({ statusCode: 500, message: "Internal Server Error" });
   }
-};
+);
 
-export const addAppointmentToPatient = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+export const addAppointmentToPatient = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
     const { patientId, ...appointmentData } = req.body;
+
+    if (!mongoose.Types.ObjectId.isValid(patientId)) {
+      return next({ statusCode: 400, message: "Invalid Patient ID format" });
+    }
     const existingPatient = await PatientModel.findById(patientId);
     if (!existingPatient) {
-      return next({ statusCode: 404, message: "PatientNot found" });
+      return next({ statusCode: 404, message: "Patient not found" });
     }
     const newAppointment = new AppointmentModel({
       ...appointmentData,
       patientId: existingPatient._id,
     });
     const savedAppointment = await newAppointment.save();
-    console.log(savedAppointment);
     if (!savedAppointment) {
       return next({ statusCode: 400, message: "Saved Appointment failed" });
     }
     res.status(201).json(savedAppointment);
-  } catch (error) {
-    return next({ statusCode: 500, message: "Internal Server Error" });
   }
-};
-export const getUnpaidAppointments = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+);
+
+export const getUnpaidAppointments = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
     const unpaidAppointments = await AppointmentModel.find({ isPaid: false });
     if (unpaidAppointments.length === 0) {
       return next({ statusCode: 404, message: "No Unpaid Appointment" });
     }
     res.status(200).json(unpaidAppointments);
-  } catch (error) {
-    return next({ statusCode: 500, message: "Internal Server Error" });
   }
-};
-export const getAppointmentsforDate = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+);
+export const getAppointmentsforDate = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
     const requestedDate = new Date(req.params.date);
     const startOfDay = new Date(
       requestedDate.getFullYear(),
@@ -118,50 +106,34 @@ export const getAppointmentsforDate = async (
     const appointments = await AppointmentModel.find({
       startTime: { $gte: startOfDay, $lte: endOfDay },
     });
-    if (!appointments || appointments.length === 0) {
+    if (appointments.length === 0) {
       return next({
         statusCode: 404,
-        message: "No Appointments for given date",
+        message: "No Appointments for the given date",
       });
     }
     res.status(200).json(appointments);
-  } catch (err) {
-    console.log(
-      `Error in getting all appointments for date ${req.params.date}`,
-      err
-    );
-    return next({ statusCode: 500, message: "Internal Server Error" });
   }
-};
-//get remaining bill of patient = bill amount of patients whose status is unpaid
-export const getRemainingBill = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+);
+export const getRemainingBill = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
     const unpaidSum = await AppointmentModel.aggregate([
       { $match: { isPaid: false } },
       { $group: { _id: null, total: { $sum: "$paymentAmount" } } },
     ]);
-    if (!unpaidSum || unpaidSum.length === 0) {
+
+    if (unpaidSum.length === 0) {
       return next({
         statusCode: 404,
         message: "No Unpaid Appointments, therefore RemainingBill = Null",
       });
     }
     res.status(200).json(`Total Remaining Bill ${unpaidSum[0].total}`);
-  } catch (error) {
-    return next({ statusCode: 500, message: "Internal Server Error" });
   }
-};
+);
 
-export const getPopularPatient = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+export const getPopularPatient = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
     const popularPatient = await AppointmentModel.aggregate([
       { $match: { isPaid: true } },
       {
@@ -183,7 +155,7 @@ export const getPopularPatient = async (
       { $unwind: { path: "$patient", preserveNullAndEmptyArrays: true } },
     ]);
 
-    if (!popularPatient || popularPatient.length === 0) {
+    if (popularPatient.length === 0) {
       return next({ statusCode: 404, message: "No Popular Patient Found" });
     }
     const popularPatientData = popularPatient[0];
@@ -191,35 +163,21 @@ export const getPopularPatient = async (
       patient: popularPatientData._id,
       totalPayment: popularPatientData.totalPayment,
     });
-  } catch (error) {
-    return next({ statusCode: 500, message: "Internal Server Error" });
   }
-};
+);
+export const getAllAppointment = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const appointments = await AppointmentModel.find();
 
-export const getAllAppointment = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    const Appointment = await AppointmentModel.find();
-
-    if (!Appointment || Appointment.length === 0) {
+    if (appointments.length === 0) {
       return next({ statusCode: 404, message: "No Appointment Available" });
     }
-    res.status(200).json(Appointment);
-  } catch (error) {
-    return next({ statusCode: 500, message: "Internal Server Error" });
+    res.status(200).json(appointments);
   }
-  (" ");
-};
+);
 
-export const getAppointmentFinancials = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
+export const getAppointmentFinancials = asyncHandler(
+  async (req: Request, res: Response, next: NextFunction) => {
     const result = await AppointmentModel.aggregate([
       {
         $group: {
@@ -230,13 +188,11 @@ export const getAppointmentFinancials = async (
           },
           paidAmount: {
             $sum: {
-              //if cond evaluate to TRUE sum is done on PaymentAmount else sum is done on 0
               $cond: [{ $eq: ["$isPaid", true] }, "$paymentAmount", 0],
             },
           },
           unpaidAmount: {
             $sum: {
-              //if cond evaluate to TRUE sum is done on PaymentAmount else sum is done on 0
               $cond: [{ $eq: ["$isPaid", false] }, "$paymentAmount", 0],
             },
           },
@@ -254,11 +210,9 @@ export const getAppointmentFinancials = async (
       },
     ]);
 
-    if (!result || result.length === 0) {
+    if (result.length === 0) {
       return next({ statusCode: 404, message: "No financial data found" });
     }
     res.status(200).json(result);
-  } catch (error) {
-    return next({ statusCode: 500, message: "Internal Server Error" });
   }
-};
+);
